Add binary_cb for binary diff data
Previously, we calculated binary diffs and presented them to the
file_cb (since a binary diff encompasses the whole file), but at
a layer that had lost the context of the files themselves (to
avoid unnecessarily loading them) and instead attempted to pull them
out of the ODB.

Instead, provide a callback just for binary data and put that data on
the `git_patch` itself.  Patch printing callbacks will either format
the binary data in the base85 delta/deflate scheme or simply print the
"these files are binary" message, depending on flags.