config: examine whole file when writing
Previously we would try to be clever when writing the configuration
file and try to stop parsing when we either found the value we were
trying to write, or when we left the section that value was in.
Regrettably, you can have another of the same section.  Thus, we
must keep reading the configuration file to ensure that we do not
write that value.