merge::workdir::dirty: make it more obvious
merge::workdir::dirty emulates a merge where someone had poorly
configured their filters (eg, checked in files with line endings
that disagreed with their CRLF settings, or changed CRLF handling
in the middle of modifications).  So actually emulate this by
changing the filter configuration instead of trying to be tricky
about emulating this.