libgit2client: a client "middleware" library
Introduce libgit2client, a client "middleware" library.  This is an
experimental set of utility functions and classes for client software
that builds on top of libgit2.

This library might contain - for example - code that invokes filters or
other tools.  This is incredibly useful to share and reuse among
consumers but should be excluded from libgit2 itself.  Users may,
understandably, not want code that executes arbitrary other commands in
libgit2 itself.